+++ src/components/button/Button.tsx
@@ -5,13 +5,18 @@ import "./Button.css";
 interface Props {
   onClick: () => void;
   className?: string;
+  title: string;
 }
 
 export const Button = (props: PropsWithChildren<Props>) => {
-  const { children, className, onClick } = props;
+  const { children, className, onClick, title } = props;
 
   return (
-    <button onClick={onClick} className={combine("button", className)}>
+    <button
+      onClick={onClick}
+      className={combine("button", className)}
+      title={title}
+    >
       {children}
     </button>
   );