+++ src/hooks/useCachedEntityCanvas.ts
@@ -1,4 +1,4 @@
-import { useMemo, useEffect } from "react";
+import { useMemo, useEffect, useState } from "react";
 import {
   entityColors,
   EntityType,
@@ -8,46 +8,70 @@ import {
   TileType,
 } from "../types/factorio";
 
+const canvasOffset: Record<EntityType | TileType, number> = {
+  roboport: -1,
+  "stone-wall": 1,
+  "refined-concrete": 1,
+  "stone-path": 1,
+  substation: 0,
+  "solar-panel": 0,
+  accumulator: 0,
+  radar: 0,
+  concrete: 1,
+};
+
 export const useCachedEntityCanvas = (
   entities: FactorioEntity[] | FactorioTile[],
   canvas: HTMLCanvasElement | null,
   type: EntityType | TileType,
   size: number,
-  offset: { x: number; y: number },
-  onComplete?: () => void
+  offset: { x: number; y: number }
 ) => {
+  const [loading, setLoading] = useState(false);
   const filteredEntities = useMemo(() => {
     return (entities as { name: string; position: FactorioPosition }[]).filter(
       (e) => e.name === type
     );
   }, [entities, type]);
 
+  const context = useMemo(() => {
+    return canvas ? canvas.getContext("2d") : null;
+  }, [canvas]);
+
+  const width = canvas?.width ?? 0;
+  const height = canvas?.height ?? 0;
+
   useEffect(() => {
-    if (canvas && filteredEntities) {
-      const context = canvas.getContext("2d");
-      const additionalOffset =
-        type === "roboport"
-          ? -1
-          : type === "stone-wall" ||
-            type === "refined-concrete" ||
-            type === "stone-path"
-          ? 1
-          : 0;
+    if (context && filteredEntities) {
+      setLoading(true);
+      new Promise((resolve) => {
+        if (context) {
+          const additionalOffset = canvasOffset[type];
+          context.clearRect(0, 0, width, height);
+          context.fillStyle = entityColors[type];
+          filteredEntities.forEach((e) => {
+            context.fillRect(
+              e.position.x + offset.x + additionalOffset,
+              e.position.y + offset.y + additionalOffset,
+              size,
+              size
+            );
+          });
+        }
 
-      if (context) {
-        context.clearRect(0, 0, canvas.width, canvas.height);
-        context.fillStyle = entityColors[type];
-        filteredEntities.forEach((e) => {
-          context.fillRect(
-            e.position.x + offset.x + additionalOffset,
-            e.position.y + offset.y + additionalOffset,
-            size,
-            size
-          );
-        });
-      }
+        resolve(true);
+      }).then(() => setLoading(false));
     }
+  }, [
+    height,
+    width,
+    context,
+    filteredEntities,
+    offset.x,
+    offset.y,
+    size,
+    type,
+  ]);
 
-    onComplete?.();
-  }, [canvas, filteredEntities, offset.x, offset.y, onComplete, size, type]);
+  return loading;
 };