+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v1.2 17.02.2022
+
+### Added
+
+- tooltips for buttons
+- discord server
+
+### Removed
+
+- code section (copy button moved to preview section)
+
 ## v1.1 14.02.2022
 
 ### Added